refactor: add price history model
